Add new typedefs to support 32/64 mixed mode.
The types int64_t, uint64_t, and double have different alignments
between common 32-bit and 64-bit platforms, and this can cause
trouble when they are used in structs which are shared between
address spaces. Introduce new typedefs for these types with
special alignment attributes to request consistent alignment.

ABI implications:

In the current tree, with packed attributes on the relevant structs,
the only impact of this change is that the alignments of jack_unique_t
and jack_time_t are increased from 4 to 8 on 32-bit platforms, so
the layouts of structs which include them as members may be affected.

The ABI impact relative to the tree before the packed attributes
were introduced is that the alignment of jack_position_t is increased
from 4 to 8 on 32-bit platforms. The layout of the struct itself is
unmodified, though the layouts of structs which include it as a
member may be affected.